todo: add landcover constraint class

+++ coe_cellular_automata/src/coe_cellular_automata/LandCoverConstraint.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.LinkedHashMap;
 
 public class LandCoverConstraint {
-	ArrayList<LinkedHashMap<String,double[]>> landCoverConstraintList = new ArrayList<LinkedHashMap<String, double[]>> ();
+	ArrayList<LinkedHashMap<String,int[]>> landCoverConstraintList = new ArrayList<LinkedHashMap<String, int[]>> ();
 	ArrayList<ArrayList<Integer>> landCoverCellList = new ArrayList<ArrayList<Integer>>();
 	
 	LandCoverConstraint(){
